Switching to per course-instance model
