+++ src/pages/courseInstance/courseInstance.sql
@@ -70,3 +70,45 @@ AND meetings.ci_year = $ciYear
 AND sections.ci_term = $ciTerm
 AND sections.ci_name = $ciName
 AND sections.ci_year = $ciYear;
+
+-- BLOCK insert_update_roster
+INSERT INTO students (
+    netid,
+    UIN,
+    ci_term,
+    ci_name,
+    ci_year,
+    lastName,
+    firstName,
+    email,
+    college,
+    major,
+    admit_term,
+    credits,
+    level
+)
+SELECT * FROM UNNEST (
+    $net_id::text[], 
+    $UIN::bigint[],
+    $ci_term::text[],
+    $ci_name::text[],
+    $ci_year::smallint[], 
+    $last_name::text[], 
+    $first_name::text[],
+    $email_address::text[],
+    $College::text[],
+    $major_name::text[],
+    $admit_term::text[],
+    $Credit::smallint[],
+    $Level::text[]
+)
+ON CONFLICT (UIN, ci_term, ci_name, ci_year) DO UPDATE
+    SET netid = excluded.netid,
+        lastName = excluded.lastName,
+        firstName = excluded.firstName,
+        email = excluded.email,
+        college = excluded.college,
+        major = excluded.major,
+        admit_term = excluded.admit_term,
+        credits = excluded.credits,
+        level = excluded.level;
