+++ utils/db_setup/student.sql
@@ -5,13 +5,13 @@ CREATE TABLE IF NOT EXISTS students (
   ci_name         varchar(80),
   ci_year         smallint,
   college         varchar(160),
-  level           smallint,
+  level           char(2),
   netid           varchar(40),
   email           varchar(80),
   firstName       varchar(80),
   lastName        varchar(80),
   preferredName   varchar(160),
-  year            smallint,
+  admit_term      varchar(20),
   major           varchar(40),
   credits         smallint,
   PRIMARY KEY (UIN, ci_term, ci_name, ci_year),
